+++ CHANGELOG.md
@@ -2,12 +2,15 @@
  All important changes in this project are stored in this file.
 
 [Semantic versioning](https://semver.org/spec/v2.0.0.html) is used.
-
 ## Unreleased - in development
 ### __Breaking Changes__
-
 ### Added
 - QuickHelp: A description of the route preference `recommended route` has been added.
+- 3D:
+  - Colored highlighting for 3D Tiles has been added.
+  - GFI MapMarker appears now on top of elevations by click.
+- Filter:
+  - Add: new parameter minZoom and maxZoom in layer configuration to make filter to react with zoom level. If the current zoom level if out of range, the filter will be deactivated.
 
 ### Changed
 
@@ -17,8 +20,12 @@
 
 ### Fixed
 - Print: The portal is no longer reloaded after pressing the Enter key in the title field. Instead, the print is sent.
-- GFI: Fixed GFI for Cesium TileFeatures by updating Cesium function name.
 - WFST-Tool can also parse Geoserver responses from DescribeFeatureType-Requests
+
+---
+## v2.35.1 - 2023-07-13
+### Fixed
+- GFI: Fixed GFI for Cesium TileFeatures by updating Cesium function name.
 
 ---
 ## v2.35.0 - 2023-07-05